Profile is setup, some logged in features added.
Updated landing index  to return the user to the home page if they're not logged in, otherwise they weill see their dashboard. 

updated home/ to redirect to to index if user is not logged in. 
updated base.html to change the log in/out button text and behvior depending on if the user is logged in or not. 
started implementing the dashboard, right now I need to add the link shortening feature.